Rename to TS & add TS support
