--- babel.config.json
@@ -1,10 +1,0 @@
-{
-  "presets": [
-    "@babel/env"
-  ],
-  "plugins": [
-    "@babel/proposal-class-properties",
-    "@babel/proposal-object-rest-spread",
-    "@babel/transform-runtime"
-  ]
-}
