+++ package.json
@@ -7,9 +7,13 @@
     "test": "test"
   },
   "scripts": {
-    "build": "swc src --out-dir lib",
+    "build": "rimraf lib && concurrently \"npm run build:src\" \"npm run build:types\"",
+    "build:watch": "rimraf lib && concurrently \"npm run build:src:watch\" \"npm run build:types:watch\"",
+    "build:src": "swc src --out-dir lib",
+    "build:src:watch": "npm run build:types -- --watch",
+    "build:types": "tsc",
     "test": "jest",
-    "test.watch": "jest --watch",
+    "test:watch": "jest --watch",
     "coverage": "jest --collectCoverage",
     "coverage-start": "jest --collectCoverage && start coverage/lcov-report/index.html",
     "eslint": "eslint src",
@@ -28,12 +32,19 @@
     "@swc/cli": "^0.1.57",
     "@swc/core": "^1.2.222",
     "@swc/jest": "^0.2.22",
+    "@types/jest": "^28.1.6",
+    "@types/lodash": "^4.14.182",
+    "@types/node": "^18.6.3",
+    "@types/uuid": "^8.3.4",
     "babel-eslint": "10.1.0",
+    "concurrently": "^7.3.0",
     "eslint": "^8.21.0",
     "eslint-config-surikat": "^3.0.0",
     "jest": "^28.1.3",
     "prettier": "^2.7.1",
-    "prettier-config-surikaterna": "^1.0.1"
+    "prettier-config-surikaterna": "^1.0.1",
+    "rimraf": "^3.0.2",
+    "typescript": "^4.7.4"
   },
   "dependencies": {
     "kuery": "^0.5.2",
