+++ src/Cursor.ts
