+++ src/merge.spec.ts
