+++ src/merge.ts
